* Changed Game::_variables to public since the GPL interpreter needs to use it and made it int instead of uint16 * Implemented variable accessing by the math evaluator * Fixed bug from previous commit (should have used && when checking for ending instructions, not ||)
svn-id: r42242

+++ engines/draci/game.cpp
@@ -96,7 +96,7 @@ Game::Game(DraciEngine *vm) : _vm(vm) {
 	file = initArchive.getFile(2);
 	unsigned int numVariables = file->_length / sizeof (int16);
 
-	_variables = new int16[numVariables];
+	_variables = new int[numVariables];
 	Common::MemoryReadStream variableData(file->_data, file->_length);
 	
 	for (i = 0; i < numVariables; ++i) {
